Revert "[PCCT] Allow customzing content rect for float action mode"
This reverts commit 2a8ef800

+++ content/public/android/java/src/org/chromium/content/browser/selection/SelectionPopupControllerImpl.java
@@ -42,6 +42,7 @@ import org.chromium.base.metrics.RecordUserAction;
 import org.chromium.base.supplier.ObservableSupplier;
 import org.chromium.base.supplier.ObservableSupplierImpl;
 import org.chromium.content.R;
+import org.chromium.content.browser.ContentApiHelperForM;
 import org.chromium.content.browser.ContentClassFactory;
 import org.chromium.content.browser.GestureListenerManagerImpl;
 import org.chromium.content.browser.PopupController;
@@ -119,7 +120,7 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
     private Context mContext;
     private WindowAndroid mWindowAndroid;
     private WebContentsImpl mWebContents;
-    private ActionMode.Callback2 mCallback;
+    private ActionMode.Callback mCallback;
     private RenderFrameHost mRenderFrameHost;
     private long mNativeSelectionPopupController;
 
@@ -340,7 +341,7 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
     }
 
     @Override
-    public void setActionModeCallback(ActionMode.Callback2 callback) {
+    public void setActionModeCallback(ActionMode.Callback callback) {
         mCallback = callback;
     }
 
@@ -410,7 +411,8 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
             boolean shouldSuggest, @MenuSourceType int sourceType,
             RenderFrameHost renderFrameHost) {
         int offsetBottom = bottom;
-        offsetBottom += handleHeight;
+        // Legacy action mode expects the selection rectangle not to include touch handle.
+        if (supportsFloatingActionMode()) offsetBottom += handleHeight;
         mSelectionRect.set(left, top, right, offsetBottom);
         mEditable = isEditable;
         mLastSelectedText = selectionText;
@@ -486,7 +488,8 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
         destroyActionModeAndKeepSelection();
 
         assert mWebContents != null;
-        ActionMode actionMode = mView.startActionMode(mCallback, ActionMode.TYPE_FLOATING);
+        ActionMode actionMode = supportsFloatingActionMode() ? startFloatingActionMode()
+                                                             : mView.startActionMode(mCallback);
         if (actionMode != null) {
             // This is to work around an LGE email issue. See crbug.com/651706 for more details.
             LGEmailActionModeWorkaroundImpl.runIfNecessary(mContext, actionMode);
@@ -495,6 +498,13 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
         mUnselectAllOnDismiss = true;
 
         if (!isActionModeValid()) clearSelection();
+    }
+
+    private ActionMode startFloatingActionMode() {
+        assert mView != null;
+        assert Build.VERSION.SDK_INT >= Build.VERSION_CODES.M;
+        ActionMode actionMode = ContentApiHelperForM.startActionMode(mView, this, mCallback);
+        return actionMode;
     }
 
     private void dismissTextHandles() {
@@ -514,6 +524,10 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
             return;
         }
 
+        if (!supportsFloatingActionMode() && !Clipboard.getInstance().canPaste()
+                && mNonSelectionCallback == null) {
+            return;
+        }
         destroyPastePopup();
         PastePopupMenu.PastePopupMenuDelegate delegate =
                 new PastePopupMenu.PastePopupMenuDelegate() {
@@ -551,8 +565,12 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
                 };
         Context windowContext = mWindowAndroid.getContext().get();
         if (windowContext == null) return;
-        mPastePopupMenu =
-                new FloatingPastePopupMenu(windowContext, mView, delegate, mNonSelectionCallback);
+        if (supportsFloatingActionMode()) {
+            mPastePopupMenu = new FloatingPastePopupMenu(
+                    windowContext, mView, delegate, mNonSelectionCallback);
+        } else {
+            mPastePopupMenu = new LegacyPastePopupMenu(windowContext, mView, delegate);
+        }
         showPastePopup();
     }
 
@@ -561,6 +579,11 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
             mPastePopupMenu.show(getSelectionRectRelativeToContainingView());
         } catch (WindowManager.BadTokenException e) {
         }
+    }
+
+    @Override
+    public boolean supportsFloatingActionMode() {
+        return Build.VERSION.SDK_INT >= Build.VERSION_CODES.M;
     }
 
     // HideablePopup implementation
@@ -603,14 +626,16 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
      * @see ActionMode#invalidateContentRect()
      */
     public void invalidateContentRect() {
-        if (isActionModeValid()) ApiHelperForM.invalidateContentRectOnActionMode(mActionMode);
+        if (supportsFloatingActionMode() && isActionModeValid()) {
+            ApiHelperForM.invalidateContentRectOnActionMode(mActionMode);
+        }
     }
 
     // WindowEventObserver
 
     @Override
     public void onWindowFocusChanged(boolean gainFocus) {
-        if (isActionModeValid()) {
+        if (supportsFloatingActionMode() && isActionModeValid()) {
             ApiHelperForM.onWindowFocusChangedOnActionMode(mActionMode, gainFocus);
         }
     }
@@ -707,12 +732,15 @@ public class SelectionPopupControllerImpl extends ActionModeCallbackHelper
     }
 
     private boolean isFloatingActionMode() {
-        return isActionModeValid()
+        return supportsFloatingActionMode() && isActionModeValid()
                 && ApiHelperForM.getActionModeType(mActionMode) == ActionMode.TYPE_FLOATING;
     }
 
     private long getDefaultHideDuration() {
-        return ApiHelperForM.getDefaultActionModeHideDuration();
+        if (supportsFloatingActionMode()) {
+            return ApiHelperForM.getDefaultActionModeHideDuration();
+        }
+        return 2000;
     }
 
     // Default handlers for action mode callbacks.
